feat: editable data sources
