--- apps/tissuumaps/src/components/panels/PointsPanel/PointsGroupsPanel.tsx
@@ -1,21 +1,0 @@
-import { type Points } from "@tissuumaps/core";
-
-import { Fieldset, FieldsetLegend } from "@/components/common/fieldset";
-import { cn } from "@/lib/utils";
-
-export type PointsGroupsPanelProps = {
-  points: Points;
-  className?: string;
-};
-
-export function PointsGroupsPanel({ className }: PointsGroupsPanelProps) {
-  return (
-    <Fieldset
-      className={cn("flex flex-col gap-y-2 border rounded-md p-2", className)}
-    >
-      <FieldsetLegend className="font-medium text-foreground">
-        Groups
-      </FieldsetLegend>
-    </Fieldset>
-  );
-}
