+++ apps/tissuumaps/src/components/panels/PointsPanel/PointsSettingsWidget.tsx
@@ -1,12 +1,12 @@
 import {
   MathUtils,
-  type Shapes,
-  defaultShapeFillColor,
-  defaultShapeFillOpacity,
-  defaultShapeFillVisibility,
-  defaultShapeStrokeColor,
-  defaultShapeStrokeOpacity,
-  defaultShapeStrokeVisibility,
+  type Points,
+  defaultPointColor,
+  defaultPointMarker,
+  defaultPointOpacity,
+  defaultPointSize,
+  defaultPointSizeUnit,
+  defaultPointVisibility,
 } from "@tissuumaps/core";
 
 import {
@@ -28,11 +28,23 @@ import {
 } from "@/components/widgets/config/ColorConfigWidget";
 import { useColorConfigWidget } from "@/components/widgets/config/ColorConfigWidget/hooks";
 import {
+  ActiveMarkerConfigValue,
+  MarkerConfigSourceToggleGroup,
+  MarkerConfigWidget,
+} from "@/components/widgets/config/MarkerConfigWidget";
+import { useMarkerConfigWidget } from "@/components/widgets/config/MarkerConfigWidget/hooks";
+import {
   ActiveOpacityConfigValue,
   OpacityConfigSourceToggleGroup,
   OpacityConfigWidget,
 } from "@/components/widgets/config/OpacityConfigWidget";
 import { useOpacityConfigWidget } from "@/components/widgets/config/OpacityConfigWidget/hooks";
+import {
+  ActiveSizeConfigValue,
+  SizeConfigSourceToggleGroup,
+  SizeConfigWidget,
+} from "@/components/widgets/config/SizeConfigWidget";
+import { useSizeConfigWidget } from "@/components/widgets/config/SizeConfigWidget/hooks";
 import {
   ActiveVisibilityConfigValue,
   VisibilityConfigSourceToggleGroup,
@@ -42,54 +54,45 @@ import { useVisibilityConfigWidget } from "@/components/widgets/config/Visibilit
 import { cn } from "@/lib/utils";
 import { useTissUUmaps } from "@/store";
 
-export type ShapesSettingsPanelProps = {
-  shapes: Shapes;
+export type PointsSettingsWidgetProps = {
+  points: Points;
   className?: string;
 };
 
-export function ShapesSettingsPanel({
-  shapes,
+export function PointsSettingsWidget({
+  points,
   className,
-}: ShapesSettingsPanelProps) {
-  const updateShapes = useTissUUmaps((state) => state.updateShapes);
+}: PointsSettingsWidgetProps) {
+  const updatePoints = useTissUUmaps((state) => state.updatePoints);
 
-  const shapeFillColorConfigWidgetAdapter = useColorConfigWidget(
-    shapes.shapeFillColor,
-    (newColorConfig) =>
-      updateShapes(shapes.id, { shapeFillColor: newColorConfig }),
-    defaultShapeFillColor,
+  const pointMarkerConfigWidgetAdapter = useMarkerConfigWidget(
+    points.pointMarker,
+    (newMarkerConfig) =>
+      updatePoints(points.id, { pointMarker: newMarkerConfig }),
+    defaultPointMarker,
   );
-  const shapeFillVisibilityConfigWidgetAdapter = useVisibilityConfigWidget(
-    shapes.shapeFillVisibility,
+  const pointSizeConfigWidgetAdapter = useSizeConfigWidget(
+    points.pointSize,
+    (newSizeConfig) => updatePoints(points.id, { pointSize: newSizeConfig }),
+    defaultPointSize,
+    defaultPointSizeUnit,
+  );
+  const pointColorConfigWidgetAdapter = useColorConfigWidget(
+    points.pointColor,
+    (newColorConfig) => updatePoints(points.id, { pointColor: newColorConfig }),
+    defaultPointColor,
+  );
+  const pointVisibilityConfigWidgetAdapter = useVisibilityConfigWidget(
+    points.pointVisibility,
     (newVisibilityConfig) =>
-      updateShapes(shapes.id, { shapeFillVisibility: newVisibilityConfig }),
-    defaultShapeFillVisibility,
+      updatePoints(points.id, { pointVisibility: newVisibilityConfig }),
+    defaultPointVisibility,
   );
-  const shapeFillOpacityConfigWidgetAdapter = useOpacityConfigWidget(
-    shapes.shapeFillOpacity,
+  const pointOpacityConfigWidgetAdapter = useOpacityConfigWidget(
+    points.pointOpacity,
     (newOpacityConfig) =>
-      updateShapes(shapes.id, { shapeFillOpacity: newOpacityConfig }),
-    defaultShapeFillOpacity,
-  );
-  const shapeStrokeColorConfigWidgetAdapter = useColorConfigWidget(
-    shapes.shapeStrokeColor,
-    (newColorConfig) =>
-      updateShapes(shapes.id, { shapeStrokeColor: newColorConfig }),
-    defaultShapeStrokeColor,
-  );
-  const shapeStrokeVisibilityConfigWidgetAdapter = useVisibilityConfigWidget(
-    shapes.shapeStrokeVisibility,
-    (newVisibilityConfig) =>
-      updateShapes(shapes.id, {
-        shapeStrokeVisibility: newVisibilityConfig,
-      }),
-    defaultShapeStrokeVisibility,
-  );
-  const shapeStrokeOpacityConfigWidgetAdapter = useOpacityConfigWidget(
-    shapes.shapeStrokeOpacity,
-    (newOpacityConfig) =>
-      updateShapes(shapes.id, { shapeStrokeOpacity: newOpacityConfig }),
-    defaultShapeStrokeOpacity,
+      updatePoints(points.id, { pointOpacity: newOpacityConfig }),
+    defaultPointOpacity,
   );
 
   return (
@@ -109,9 +112,9 @@ export function ShapesSettingsPanel({
             <Field>
               <FieldLabel>Name</FieldLabel>
               <Input
-                value={shapes.name}
+                value={points.name}
                 onChange={(event) =>
-                  updateShapes(shapes.id, { name: event.target.value })
+                  updatePoints(points.id, { name: event.target.value })
                 }
               />
             </Field>
@@ -119,12 +122,12 @@ export function ShapesSettingsPanel({
               <FieldLabel>Visibility</FieldLabel>
               <div className="flex flex-row items-center gap-x-2">
                 <Switch
-                  checked={shapes.visibility}
+                  checked={points.visibility}
                   onCheckedChange={(checked) =>
-                    updateShapes(shapes.id, { visibility: checked })
+                    updatePoints(points.id, { visibility: checked })
                   }
                 />
-                {shapes.visibility ? "Visible" : "Hidden"}
+                {points.visibility ? "Visible" : "Hidden"}
               </div>
             </Field>
             <Field>
@@ -135,10 +138,10 @@ export function ShapesSettingsPanel({
                 step={0.01}
                 min={0}
                 max={1}
-                value={shapes.opacity}
+                value={points.opacity}
                 onChange={(event) => {
                   if (event.target.value !== "") {
-                    updateShapes(shapes.id, {
+                    updatePoints(points.id, {
                       opacity: MathUtils.clamp(
                         parseFloat(event.target.value),
                         0,
@@ -149,122 +152,118 @@ export function ShapesSettingsPanel({
                 }}
               />
             </Field>
+            <Field>
+              <FieldLabel>Point size factor</FieldLabel>
+              <Input
+                type="number"
+                inputMode="decimal"
+                step={0.1}
+                min={0}
+                value={points.pointSizeFactor}
+                onChange={(event) => {
+                  if (event.target.value !== "") {
+                    updatePoints(points.id, {
+                      pointSizeFactor: Math.max(
+                        0,
+                        parseFloat(event.target.value),
+                      ),
+                    });
+                  }
+                }}
+              />
+            </Field>
           </AccordionPanel>
         </AccordionItem>
-        {/* Shape fill color */}
-        <AccordionItem value="shapeFillColor">
+        {/* Point marker */}
+        <AccordionItem value="pointMarker">
           <AccordionHeader>
             <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Fill color</AccordionTrigger>
+            <AccordionTrigger>Point marker</AccordionTrigger>
+            <ActiveMarkerConfigValue
+              adapter={pointMarkerConfigWidgetAdapter}
+              className="ml-auto text-sm text-slate-600 dark:text-slate-400"
+            />
+          </AccordionHeader>
+          <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
+            <MarkerConfigSourceToggleGroup
+              adapter={pointMarkerConfigWidgetAdapter}
+              className="border rounded"
+            />
+            <MarkerConfigWidget adapter={pointMarkerConfigWidgetAdapter} />
+          </AccordionPanel>
+        </AccordionItem>
+        {/* Point size */}
+        <AccordionItem value="pointSize">
+          <AccordionHeader>
+            <AccordionTriggerRightDownIcon />
+            <AccordionTrigger>Point size</AccordionTrigger>
+            <ActiveSizeConfigValue
+              adapter={pointSizeConfigWidgetAdapter}
+              className="ml-auto text-sm text-slate-600 dark:text-slate-400"
+            />
+          </AccordionHeader>
+          <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
+            <SizeConfigSourceToggleGroup
+              adapter={pointSizeConfigWidgetAdapter}
+              className="border rounded"
+            />
+            <SizeConfigWidget adapter={pointSizeConfigWidgetAdapter} />
+          </AccordionPanel>
+        </AccordionItem>
+        {/* Point color */}
+        <AccordionItem value="pointColor">
+          <AccordionHeader>
+            <AccordionTriggerRightDownIcon />
+            <AccordionTrigger>Point color</AccordionTrigger>
             <ActiveColorConfigValue
-              adapter={shapeFillColorConfigWidgetAdapter}
+              adapter={pointColorConfigWidgetAdapter}
               className="ml-auto text-sm text-slate-600 dark:text-slate-400"
             />
           </AccordionHeader>
           <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
             <ColorConfigSourceToggleGroup
-              adapter={shapeFillColorConfigWidgetAdapter}
+              adapter={pointColorConfigWidgetAdapter}
               className="border rounded"
             />
-            <ColorConfigWidget adapter={shapeFillColorConfigWidgetAdapter} />
+            <ColorConfigWidget adapter={pointColorConfigWidgetAdapter} />
           </AccordionPanel>
         </AccordionItem>
-        {/* Shape fill visibility */}
-        <AccordionItem value="shapeFillVisibility">
+        {/* Point visibility */}
+        <AccordionItem value="pointVisibility">
           <AccordionHeader>
             <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Fill visibility</AccordionTrigger>
+            <AccordionTrigger>Point visibility</AccordionTrigger>
             <ActiveVisibilityConfigValue
-              adapter={shapeFillVisibilityConfigWidgetAdapter}
+              adapter={pointVisibilityConfigWidgetAdapter}
               className="ml-auto text-sm text-slate-600 dark:text-slate-400"
             />
           </AccordionHeader>
           <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
             <VisibilityConfigSourceToggleGroup
-              adapter={shapeFillVisibilityConfigWidgetAdapter}
+              adapter={pointVisibilityConfigWidgetAdapter}
               className="border rounded"
             />
             <VisibilityConfigWidget
-              adapter={shapeFillVisibilityConfigWidgetAdapter}
+              adapter={pointVisibilityConfigWidgetAdapter}
             />
           </AccordionPanel>
         </AccordionItem>
-        {/* Shape fill opacity */}
-        <AccordionItem value="shapeFillOpacity">
+        {/* Point opacity */}
+        <AccordionItem value="pointOpacity">
           <AccordionHeader>
             <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Fill opacity</AccordionTrigger>
+            <AccordionTrigger>Point opacity</AccordionTrigger>
             <ActiveOpacityConfigValue
-              adapter={shapeFillOpacityConfigWidgetAdapter}
+              adapter={pointOpacityConfigWidgetAdapter}
               className="ml-auto text-sm text-slate-600 dark:text-slate-400"
             />
           </AccordionHeader>
           <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
             <OpacityConfigSourceToggleGroup
-              adapter={shapeFillOpacityConfigWidgetAdapter}
+              adapter={pointOpacityConfigWidgetAdapter}
               className="border rounded"
             />
-            <OpacityConfigWidget
-              adapter={shapeFillOpacityConfigWidgetAdapter}
-            />
-          </AccordionPanel>
-        </AccordionItem>
-        {/* Shape stroke color */}
-        <AccordionItem value="shapeStrokeColor">
-          <AccordionHeader>
-            <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Stroke color</AccordionTrigger>
-            <ActiveColorConfigValue
-              adapter={shapeStrokeColorConfigWidgetAdapter}
-              className="ml-auto text-sm text-slate-600 dark:text-slate-400"
-            />
-          </AccordionHeader>
-          <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
-            <ColorConfigSourceToggleGroup
-              adapter={shapeStrokeColorConfigWidgetAdapter}
-              className="border rounded"
-            />
-            <ColorConfigWidget adapter={shapeStrokeColorConfigWidgetAdapter} />
-          </AccordionPanel>
-        </AccordionItem>
-        {/* Shape stroke visibility */}
-        <AccordionItem value="shapeStrokeVisibility">
-          <AccordionHeader>
-            <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Stroke visibility</AccordionTrigger>
-            <ActiveVisibilityConfigValue
-              adapter={shapeStrokeVisibilityConfigWidgetAdapter}
-              className="ml-auto text-sm text-slate-600 dark:text-slate-400"
-            />
-          </AccordionHeader>
-          <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
-            <VisibilityConfigSourceToggleGroup
-              adapter={shapeStrokeVisibilityConfigWidgetAdapter}
-              className="border rounded"
-            />
-            <VisibilityConfigWidget
-              adapter={shapeStrokeVisibilityConfigWidgetAdapter}
-            />
-          </AccordionPanel>
-        </AccordionItem>
-        {/* Shape stroke opacity */}
-        <AccordionItem value="shapeStrokeOpacity">
-          <AccordionHeader>
-            <AccordionTriggerRightDownIcon />
-            <AccordionTrigger>Stroke opacity</AccordionTrigger>
-            <ActiveOpacityConfigValue
-              adapter={shapeStrokeOpacityConfigWidgetAdapter}
-              className="ml-auto text-sm text-slate-600 dark:text-slate-400"
-            />
-          </AccordionHeader>
-          <AccordionPanel className="flex flex-col p-2 pl-6 pb-4 gap-2">
-            <OpacityConfigSourceToggleGroup
-              adapter={shapeStrokeOpacityConfigWidgetAdapter}
-              className="border rounded"
-            />
-            <OpacityConfigWidget
-              adapter={shapeStrokeOpacityConfigWidgetAdapter}
-            />
+            <OpacityConfigWidget adapter={pointOpacityConfigWidgetAdapter} />
           </AccordionPanel>
         </AccordionItem>
       </Accordion>
