+++ apps/tissuumaps/src/components/widgets/DataSourceWidget/renderers.ts
@@ -1,23 +1,11 @@
 import {
   and,
-  isBooleanControl,
   isControl,
-  isEnumControl,
-  isIntegerControl,
-  isNumberControl,
   isPrimitiveArrayControl,
-  isStringControl,
   rankWith,
-  scopeEndIs,
   uiTypeIs,
 } from "@jsonforms/core";
 
-import { BooleanCell } from "./cells/BooleanCell";
-import { IntegerCell } from "./cells/IntegerCell";
-import { NumberCell } from "./cells/NumberCell";
-import { StringEnumCell } from "./cells/StringEnumCell";
-import { TableEnumCell } from "./cells/TableEnumCell";
-import { TextCell } from "./cells/TextCell";
 import { ArrayControl } from "./controls/ArrayControl";
 import { InputControl } from "./controls/InputControl";
 import { LabelControl } from "./controls/LabelControl";
@@ -37,20 +25,5 @@ export const renderers = [
   {
     tester: rankWith(2, and(isControl, isPrimitiveArrayControl)),
     renderer: ArrayControl,
-  },
-];
-
-export const cells = [
-  { tester: rankWith(1, isBooleanControl), cell: BooleanCell },
-  { tester: rankWith(1, isIntegerControl), cell: IntegerCell },
-  { tester: rankWith(1, isNumberControl), cell: NumberCell },
-  { tester: rankWith(1, isStringControl), cell: TextCell },
-  {
-    tester: rankWith(2, and(isStringControl, isEnumControl)),
-    cell: StringEnumCell,
-  },
-  {
-    tester: rankWith(3, and(isStringControl, scopeEndIs("table"))),
-    cell: TableEnumCell,
   },
 ];
