+++ packages/@tissuumaps-storage/src/parquet/ParquetTableDataProvider.ts
@@ -33,7 +33,7 @@ export class ParquetTableDataProvider implements TableDataProvider<
     required: ["url"], // TODO ... or path
   };
 
-  readonly uiSchema = {
+  readonly uischema = {
     type: "VerticalLayout",
     elements: [
       {
